Add warning to LV volumes computation when (likely) multi-class masks are provided (#158)
The `compute_left_ventricle_volumes` function expects binary masks of the left ventricle (LV), but won't error out on segmentation with arbitrary labels (instead treating any "truthy" value as part of the binary mask). This lead users on multiple occasions to provide multi-class segmentation masks (w/o the LV being extracted first), which lead to obviously wrong results but w/o any errors.

Therefore, this warning is being added to flag the user-error quicker in the future, w/o outright raising an error, in case some users are aware of the masking behavior and rely on it.