Add config option for `torchinfo.summary`'s `mode` parameter
In some cases, some layers (e.g. `TransformerEncoderLayer`) can use fast execution paths that bypass their submodule, failing to register the submodules for `torchinfo`.

In those cases, it can be necessary to use `train` mode to force the execution of the submodules so that they can be picked up by `torchinfo`.